CAUSEWAY-3390: fixed integ test (?); also noticed other repositories ...
... requiring the same treatment
