+++ extensions/core/commandlog/applib/src/main/java/org/apache/causeway/extensions/commandlog/applib/dom/BackgroundService.java
@@ -117,7 +117,7 @@ public class BackgroundService {
     @Service
     public static class PersistCommandExecutorService implements ExecutorService {
 
-        @Inject CommandLogEntryRepository<? extends CommandLogEntry> commandLogEntryRepository;
+        @Inject CommandLogEntryRepositoryAbstract<? extends CommandLogEntry> commandLogEntryRepository;
 
         private final static JavaSqlJaxbAdapters.TimestampToXMLGregorianCalendarAdapter gregorianCalendarAdapter  = new JavaSqlJaxbAdapters.TimestampToXMLGregorianCalendarAdapter();;
