+++ extensions/core/commandlog/applib/src/main/java/org/apache/causeway/extensions/commandlog/applib/dom/CommandLogEntry.java
@@ -382,7 +382,8 @@ implements Comparable<CommandLogEntry>, DomainChangeRecord, HasCommandDto {
         return _Casts.uncheckedCast(commandLogEntry);
     }
 
-    @Inject CommandLogEntryRepository<? extends CommandLogEntry> commandLogEntryRepository;
+    @Inject
+    CommandLogEntryRepositoryAbstract<? extends CommandLogEntry> commandLogEntryRepository;
 
 
     @Property(
