+++ extensions/core/commandlog/applib/src/main/java/org/apache/causeway/extensions/commandlog/applib/dom/mixins/CommandLogEntry_siblingCommands.java
@@ -27,7 +27,7 @@ import org.apache.causeway.applib.annotation.CollectionLayout;
 import org.apache.causeway.applib.annotation.MemberSupport;
 import org.apache.causeway.extensions.commandlog.applib.CausewayModuleExtCommandLogApplib;
 import org.apache.causeway.extensions.commandlog.applib.dom.CommandLogEntry;
-import org.apache.causeway.extensions.commandlog.applib.dom.CommandLogEntryRepository;
+import org.apache.causeway.extensions.commandlog.applib.dom.CommandLogEntryRepositoryAbstract;
 
 import lombok.RequiredArgsConstructor;
 import lombok.val;
@@ -65,6 +65,6 @@ public class CommandLogEntry_siblingCommands {
 
 
     @javax.inject.Inject
-    private CommandLogEntryRepository<? extends CommandLogEntry> commandLogEntryRepository;
+    private CommandLogEntryRepositoryAbstract<? extends CommandLogEntry> commandLogEntryRepository;
 
 }
