+++ extensions/core/commandlog/applib/src/main/java/org/apache/causeway/extensions/commandlog/applib/fakescheduler/FakeScheduler.java
@@ -37,7 +37,7 @@ import org.apache.causeway.commons.internal.base._NullSafe;
 import org.apache.causeway.commons.internal.concurrent._ConcurrentContext;
 import org.apache.causeway.commons.internal.concurrent._ConcurrentTaskList;
 import org.apache.causeway.extensions.commandlog.applib.dom.CommandLogEntry;
-import org.apache.causeway.extensions.commandlog.applib.dom.CommandLogEntryRepository;
+import org.apache.causeway.extensions.commandlog.applib.dom.CommandLogEntryRepositoryAbstract;
 import org.apache.causeway.schema.cmd.v2.CommandDto;
 
 import lombok.Builder;
@@ -151,7 +151,8 @@ public class FakeScheduler {
         );
     }
 
-    @Inject CommandLogEntryRepository<? extends CommandLogEntry> commandLogEntryRepository;
+    @Inject
+    CommandLogEntryRepositoryAbstract<? extends CommandLogEntry> commandLogEntryRepository;
     @Inject CommandExecutorService commandExecutorService;
     @Inject TransactionService transactionService;
     @Inject InteractionService interactionService;
