+++ extensions/core/executionlog/applib/src/main/java/org/apache/causeway/extensions/executionlog/applib/app/ExecutionLogMenu.java
@@ -25,6 +25,7 @@ import java.util.List;
 import javax.inject.Inject;
 import javax.inject.Named;
 
+import org.apache.causeway.extensions.executionlog.applib.dom.ExecutionLogEntryRepositoryAbstract;
 import org.springframework.lang.Nullable;
 
 import org.apache.causeway.applib.annotation.Action;
@@ -40,7 +41,6 @@ import org.apache.causeway.applib.annotation.SemanticsOf;
 import org.apache.causeway.applib.services.clock.ClockService;
 import org.apache.causeway.extensions.executionlog.applib.CausewayModuleExtExecutionLogApplib;
 import org.apache.causeway.extensions.executionlog.applib.dom.ExecutionLogEntry;
-import org.apache.causeway.extensions.executionlog.applib.dom.ExecutionLogEntryRepository;
 
 import lombok.RequiredArgsConstructor;
 
@@ -67,7 +67,7 @@ public class ExecutionLogMenu {
             extends CausewayModuleExtExecutionLogApplib.ActionDomainEvent<T> { }
 
 
-    final ExecutionLogEntryRepository<? extends ExecutionLogEntry> executionLogEntryRepository;
+    final ExecutionLogEntryRepositoryAbstract<? extends ExecutionLogEntry> executionLogEntryRepository;
     final ClockService clockService;
