+++ extensions/core/executionlog/persistence-jpa/src/main/java/org/apache/causeway/extensions/executionlog/jpa/dom/ExecutionLogEntryRepository.java
@@ -21,6 +21,8 @@ package org.apache.causeway.extensions.executionlog.jpa.dom;
 import javax.inject.Named;
 import javax.inject.Provider;
 
+import org.apache.causeway.extensions.executionlog.applib.dom.ExecutionLogEntryRepositoryAbstract;
+
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
@@ -36,7 +38,7 @@ import lombok.Builder;
 @javax.annotation.Priority(PriorityPrecedence.MIDPOINT)
 @Qualifier("Jdo")
 public class ExecutionLogEntryRepository
-extends org.apache.causeway.extensions.executionlog.applib.dom.ExecutionLogEntryRepository<ExecutionLogEntry> {
+extends ExecutionLogEntryRepositoryAbstract<ExecutionLogEntry> {
 
     public static final String LOGICAL_TYPE_NAME = CausewayModuleExtExecutionLogPersistenceJpa.NAMESPACE + ".ExecutionLogEntryRepository";
