+++ extensions/core/executionoutbox/applib/src/test/java/org/apache/causeway/extensions/executionoutbox/applib/integtest/ExecutionOutbox_IntegTestAbstract.java
@@ -41,7 +41,7 @@ import org.apache.causeway.applib.services.user.UserMemento;
 import org.apache.causeway.applib.services.wrapper.WrapperFactory;
 import org.apache.causeway.core.config.presets.CausewayPresets;
 import org.apache.causeway.extensions.executionoutbox.applib.dom.ExecutionOutboxEntry;
-import org.apache.causeway.extensions.executionoutbox.applib.dom.ExecutionOutboxEntryRepository;
+import org.apache.causeway.extensions.executionoutbox.applib.dom.ExecutionOutboxEntryRepositoryAbstract;
 import org.apache.causeway.extensions.executionoutbox.applib.integtest.model.Counter;
 import org.apache.causeway.extensions.executionoutbox.applib.integtest.model.CounterRepository;
 import org.apache.causeway.extensions.executionoutbox.applib.integtest.model.Counter_bumpUsingMixin;
@@ -283,7 +283,8 @@ public abstract class ExecutionOutbox_IntegTestAbstract extends CausewayIntegrat
 
     }
 
-    @Inject ExecutionOutboxEntryRepository<? extends ExecutionOutboxEntry> executionOutboxEntryRepository;
+    @Inject
+    ExecutionOutboxEntryRepositoryAbstract<? extends ExecutionOutboxEntry> executionOutboxEntryRepository;
     @Inject SudoService sudoService;
     @Inject ClockService clockService;
     @Inject InteractionService interactionService;
