+++ extensions/core/executionrepublisher/applib/src/main/java/org/apache/causeway/extensions/executionrepublisher/applib/contributions/ExecutionLogEntry_copyToOutbox.java
@@ -31,7 +31,7 @@ import org.apache.causeway.applib.annotation.SemanticsOf;
 import org.apache.causeway.extensions.executionlog.applib.dom.ExecutionLogEntry;
 import org.apache.causeway.extensions.executionlog.applib.dom.ExecutionLogEntryType;
 import org.apache.causeway.extensions.executionoutbox.applib.dom.ExecutionOutboxEntry;
-import org.apache.causeway.extensions.executionoutbox.applib.dom.ExecutionOutboxEntryRepository;
+import org.apache.causeway.extensions.executionoutbox.applib.dom.ExecutionOutboxEntryRepositoryAbstract;
 import org.apache.causeway.extensions.executionoutbox.applib.dom.ExecutionOutboxEntryType;
 import org.apache.causeway.extensions.executionrepublisher.applib.CausewayModuleExtExecutionRepublisherApplib;
 
@@ -91,6 +91,7 @@ public class ExecutionLogEntry_copyToOutbox {
                 : ExecutionOutboxEntryType.PROPERTY_EDIT;
     }
 
-    @Inject ExecutionOutboxEntryRepository<? extends ExecutionOutboxEntry> outboxRepository;
+    @Inject
+    ExecutionOutboxEntryRepositoryAbstract<? extends ExecutionOutboxEntry> outboxRepository;
 
 }
