+++ extensions/security/sessionlog/applib/src/main/java/org/apache/causeway/extensions/sessionlog/applib/app/SessionLogMenu.java
@@ -38,7 +38,7 @@ import org.apache.causeway.applib.annotation.Publishing;
 import org.apache.causeway.applib.annotation.SemanticsOf;
 import org.apache.causeway.extensions.sessionlog.applib.CausewayModuleExtSessionLogApplib;
 import org.apache.causeway.extensions.sessionlog.applib.dom.SessionLogEntry;
-import org.apache.causeway.extensions.sessionlog.applib.dom.SessionLogEntryRepository;
+import org.apache.causeway.extensions.sessionlog.applib.dom.SessionLogEntryRepositoryAbstract;
 
 import lombok.RequiredArgsConstructor;
 
@@ -60,7 +60,7 @@ public class SessionLogMenu {
 
     static final String LOGICAL_TYPE_NAME = CausewayModuleExtSessionLogApplib.NAMESPACE + ".SessionLogMenu";
 
-    final SessionLogEntryRepository<? extends SessionLogEntry> sessionLogEntryRepository;
+    final SessionLogEntryRepositoryAbstract<? extends SessionLogEntry> sessionLogEntryRepository;
 
     public static abstract class ActionDomainEvent<T> extends CausewayModuleExtSessionLogApplib.ActionDomainEvent<T> { }
