+++ extensions/security/sessionlog/applib/src/main/java/org/apache/causeway/extensions/sessionlog/applib/contributions/HasUsername_recentSessionsForUser.java
@@ -33,7 +33,7 @@ import org.apache.causeway.applib.annotation.SemanticsOf;
 import org.apache.causeway.applib.mixins.security.HasUsername;
 import org.apache.causeway.extensions.sessionlog.applib.CausewayModuleExtSessionLogApplib;
 import org.apache.causeway.extensions.sessionlog.applib.dom.SessionLogEntry;
-import org.apache.causeway.extensions.sessionlog.applib.dom.SessionLogEntryRepository;
+import org.apache.causeway.extensions.sessionlog.applib.dom.SessionLogEntryRepositoryAbstract;
 
 import lombok.RequiredArgsConstructor;
 
@@ -70,6 +70,7 @@ public class HasUsername_recentSessionsForUser {
         return hasUsername == null || hasUsername.getUsername() == null;
     }
 
-    @Inject SessionLogEntryRepository<? extends SessionLogEntry> sessionLogEntryRepository;
+    @Inject
+    SessionLogEntryRepositoryAbstract<? extends SessionLogEntry> sessionLogEntryRepository;
 
 }
