+++ extensions/security/sessionlog/applib/src/main/java/org/apache/causeway/extensions/sessionlog/applib/dom/SessionLogEntryRepository.java
@@ -43,163 +43,52 @@ import lombok.val;
  *
  * @since 2.0 {@index}
  */
-public abstract class SessionLogEntryRepository<E extends SessionLogEntry> {
+public interface SessionLogEntryRepository {
 
-    @Inject RepositoryService repositoryService;
-    @Inject TransactionService transactionService;
-    @Inject FactoryService factoryService;
-    @Inject CausewaySystemEnvironment causewaySystemEnvironment;
+    void logoutAllSessions(final Timestamp logoutTimestamp);
 
-    private final Class<E> sessionLogEntryClass;
-
-    protected SessionLogEntryRepository(@NonNull Class<E> sessionLogEntryClass) {
-        this.sessionLogEntryClass = sessionLogEntryClass;
-    }
-
-    public void logoutAllSessions(final Timestamp logoutTimestamp) {
-        val allSessions = repositoryService.allMatches(
-                Query.named(sessionLogEntryClass, SessionLogEntry.Nq.FIND_ACTIVE_SESSIONS));
-        for (val activeEntry : allSessions) {
-            activeEntry.setCausedBy(SessionSubscriber.CausedBy.RESTART);
-            activeEntry.setLogoutTimestamp(logoutTimestamp);
-        }
-        transactionService.flushTransaction();
-    }
-
-    public SessionLogEntry create(
+    SessionLogEntry create(
             final String username,
             final UUID sessionGuid,
             final String httpSessionId,
             final SessionSubscriber.CausedBy causedBy,
-            final Timestamp timestamp) {
-        E entry = factoryService.detachedEntity(sessionLogEntryClass);
-        entry.setUsername(username);
-        entry.setSessionGuid(sessionGuid);
-        entry.setHttpSessionId(httpSessionId);
-        entry.setCausedBy(causedBy);
-        entry.setLoginTimestamp(timestamp);
-        return repositoryService.persistAndFlush(entry);
-    }
+            final Timestamp timestamp);
 
 
-    public Optional<E> findBySessionGuid(final UUID sessionGuid) {
-        return repositoryService.firstMatch(
-                Query.named(sessionLogEntryClass,  SessionLogEntry.Nq.FIND_BY_SESSION_GUID)
-                     .withParameter("sessionGuid", sessionGuid));
-    }
+    Optional<SessionLogEntry> findBySessionGuid(final UUID sessionGuid);
 
 
-    public Optional<E> findByHttpSessionId(final String httpSessionId) {
-        return repositoryService.firstMatch(
-                Query.named(sessionLogEntryClass,  SessionLogEntry.Nq.FIND_BY_HTTP_SESSION_ID)
-                     .withParameter("httpSessionId", httpSessionId));
-    }
+    Optional<SessionLogEntry> findByHttpSessionId(final String httpSessionId);
 
 
-    public List<E> findByUsername(final String username) {
-        return repositoryService.allMatches(
-                Query.named(sessionLogEntryClass, SessionLogEntry.Nq.FIND_BY_USERNAME)
-                     .withParameter("username", username));
-    }
+    List<SessionLogEntry> findByUsername(final String username);
 
 
-    public List<E> findByUsernameAndFromAndTo(
+    List<SessionLogEntry> findByUsernameAndFromAndTo(
             final String username,
             final LocalDate from,
-            final LocalDate to) {
-        val fromTs = toTimestampStartOfDayWithOffset(from, 0);
-        val toTs = toTimestampStartOfDayWithOffset(to, 1);
-
-        final Query<E> query;
-        if(from != null) {
-            if(to != null) {
-                query = Query.named(sessionLogEntryClass, SessionLogEntry.Nq.FIND_BY_USERNAME_AND_TIMESTAMP_BETWEEN)
-                        .withParameter("username", username)
-                        .withParameter("from", fromTs)
-                        .withParameter("to", toTs);
-            } else {
-                query = Query.named(sessionLogEntryClass, SessionLogEntry.Nq.FIND_BY_USERNAME_AND_TIMESTAMP_AFTER)
-                        .withParameter("username", username)
-                        .withParameter("from", fromTs);
-            }
-        } else {
-            if(to != null) {
-                query = Query.named(sessionLogEntryClass, SessionLogEntry.Nq.FIND_BY_USERNAME_AND_TIMESTAMP_BEFORE)
-                        .withParameter("username", username)
-                        .withParameter("to", toTs);
-            } else {
-                query = Query.named(sessionLogEntryClass, SessionLogEntry.Nq.FIND_BY_USERNAME)
-                        .withParameter("username", username);
-            }
-        }
-        return repositoryService.allMatches(query);
-    }
+            final LocalDate to);
 
 
-    public List<E> findByFromAndTo(
+    List<SessionLogEntry> findByFromAndTo(
             final LocalDate from,
-            final LocalDate to) {
-        val fromTs = toTimestampStartOfDayWithOffset(from, 0);
-        val toTs = toTimestampStartOfDayWithOffset(to, 1);
-
-        final Query<E> query;
-        if(from != null) {
-            if(to != null) {
-                query = Query.named(sessionLogEntryClass, SessionLogEntry.Nq.FIND_BY_TIMESTAMP_BETWEEN)
-                        .withParameter("from", fromTs)
-                        .withParameter("to", toTs);
-            } else {
-                query = Query.named(sessionLogEntryClass, SessionLogEntry.Nq.FIND_BY_TIMESTAMP_AFTER)
-                        .withParameter("from", fromTs);
-            }
-        } else {
-            if(to != null) {
-                query = Query.named(sessionLogEntryClass, SessionLogEntry.Nq.FIND_BY_TIMESTAMP_BEFORE)
-                        .withParameter("to", toTs);
-            } else {
-                query = Query.named(sessionLogEntryClass, SessionLogEntry.Nq.FIND);
-            }
-        }
-        return repositoryService.allMatches(query);
-    }
+            final LocalDate to);
 
 
-    public List<E> findByUsernameAndStrictlyBefore(
+    List<SessionLogEntry> findByUsernameAndStrictlyBefore(
             final String username,
-            final Timestamp from) {
-
-        return repositoryService.allMatches(
-                Query.named(sessionLogEntryClass, SessionLogEntry.Nq.FIND_BY_USERNAME_AND_TIMESTAMP_STRICTLY_BEFORE)
-                    .withParameter("username", username)
-                    .withParameter("from", from));
-    }
+            final Timestamp from);
 
 
-    public List<E> findByUsernameAndStrictlyAfter(
+    List<SessionLogEntry> findByUsernameAndStrictlyAfter(
             final String username,
-            final Timestamp from) {
-        return repositoryService.allMatches(
-                Query.named(sessionLogEntryClass, SessionLogEntry.Nq.FIND_BY_USERNAME_AND_TIMESTAMP_STRICTLY_AFTER)
-                    .withParameter("username", username)
-                    .withParameter("from", from));
-    }
+            final Timestamp from);
 
 
-
-    public List<E> findActiveSessions() {
-        return repositoryService.allMatches(
-                Query.named(sessionLogEntryClass, SessionLogEntry.Nq.FIND_ACTIVE_SESSIONS));
-    }
+    List<SessionLogEntry> findActiveSessions();
 
 
-
-    public List<E> findRecentByUsername(final String username) {
-        return repositoryService.allMatches(
-                Query.named(sessionLogEntryClass, SessionLogEntry.Nq.FIND_RECENT_BY_USERNAME)
-                        .withParameter("username", username)
-                        .withLimit(10));
-
-    }
+    List<SessionLogEntry> findRecentByUsername(final String username);
 
     private static Timestamp toTimestampStartOfDayWithOffset(final LocalDate dt, final int daysOffset) {
         return dt != null
@@ -211,21 +100,11 @@ public abstract class SessionLogEntryRepository<E extends SessionLogEntry> {
     /**
      * for testing purposes only
      */
-    public List<E> findAll() {
-        if (causewaySystemEnvironment.getDeploymentType().isProduction()) {
-            throw new IllegalStateException("Cannot removeAll in production systems");
-        }
-        return repositoryService.allInstances(sessionLogEntryClass);
-    }
+    List<SessionLogEntry> findAll();
 
     /**
      * for testing purposes only
      */
-    public void removeAll() {
-        if (causewaySystemEnvironment.getDeploymentType().isProduction()) {
-            throw new IllegalStateException("Cannot removeAll in production systems");
-        }
-        repositoryService.removeAll(sessionLogEntryClass);
-    }
+    void removeAll();
 
 }
